requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
supabase>=2.4.0
langchain-huggingface>=0.0.3
langchain-community>=0.2.0
sentence-transformers>=3.0.0
torch>=2.2.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
httpx>=0.27.0
